#42 BeanMapper.map function refactored into its constituent strategies, ie i) mapping a collection, ii) mapping to a class which must be constructed and iii) mapping to an instance. The various strategies are determined by checking key values in the configuration; a) container class, b) target class and c) target. If none of those is set, an error is assumed.
Override configuration has the concept of a nullable field, which allows a tri-state for the field; i) if it is set, but empty (ie, null), its result will be null, ii) if it is set with a proper value, that will be the legal value returned and iii) if it is null, it will fall back to its parent configuration. The NullableField concept must be propagated to targetClass and target as well. It responds to unsetX, which sets it to null, and therefore triggers the fallback.

Added a couple of tests for BeanMapperBuilder, including the expected exceptions on calling illegal core configuration operations.

Two types of config can be requested from the BeanMapper. The first one is the current config, which is either the current Override config, or the Core Config wrapped in a new Override config. The second one is the wrap config, which always wraps in a new Override config. Using the first option allows usage without constantly wrapped in new configuration, therefore being more efficient.

MapCollectionStrategy updated to reflect the new approach. In the loop to map the individual elements of a collection, it will use a beanmapper wrapped with a new override config, to ascertain that collection class will be reset to a NullableField containing null (therefore being null, and not triggering the map collection strategy).

AbstractConfiguration added for generic methods applying to both Core and Override configuations. Must be extended, probably more reuse possible.

The addition of bean converters split in a custom and default part. This might be useful later on, when converters are set separately in an override config (not currently supported).