#66 Introduces the option to mark a field as a recipient for the parent Object. This is especially handy when the ORM (eg, Hibernate) mappings are based on children needing a reference to the parent. A child can then set its reference to the parent during the mapping phase, which reduces manual mapping and -- as an added bonus -- gives you the option to validate before the controller method is hit.
One caveat I'm not too happy with; the Converter interface is changed to make sure the parent can be passed. The reason for this is that mapping a collection is handled by a converter. This converter then calls the BeanMapper.map method. On this last call it needs to pass the parent. Therefore the converter must know what the parent is. Perhaps there is a better solution for this?

The OverrideConfiguration of BeanMapper accepts the parent as a possible field. If the field cannot be found on the current override config, it will scan up the tree until it finds a legit value or hits the Core config.