#66 Instead of passing the parent to BeanConverter.convert, now the BeanMapper is passed instead. Any BeanConverter can decide to make use of this instance or not. When the parent is required, it can be fetched from the Configuration of the passed BeanMapper.
This has a huge advantage over the old way of doing things. Previously, the BeanMapper was bound to the Converter at the moment that the BeanMapper was instantiated. The downside of this approach is that the BeanMapper with its CoreConfiguration is always used. However, when you are further down the line during the mapping process, you might want to call on an enhanced BeanMapper (with one or more layers of OverrideConfiguration canvassed on it). This approach does exactly that. The called of the BeanConverter that deals with overridden configs on a BeanMapper instance and passes those directly to the BeanConverter.

BeanMapperAware has now become deprecated. During construction, the BeanMapper is not required for registering the BeanConverters. All implementing BeanConverters can just add the single BeanMapper parameter to the convert argument list and remove BeanMapperAware as an implemented interface.