Added targetCollectionType to @BeanCollection annotation
It is now possible to define a specific target type of
collection in the @BeanCollection annotation. Previously
in the case of a CONSTRUCT operation a TreeSet was always
created, but this doesn't work for classes that do not
implement Comparable. This way the user has a choice what
target collection type to choose.