Optional strict handling of BeanMapper mappings
Issue #75. Added ability to add pairs of classes which must be checked as soon as a BeanMapper instance is created.

As soon as a BeanMapper instance is created (method build on BeanMapperBuilder), it will validate its
sets of Bean pairs. The strict side will be verified so that all properties contain matching properties
on the other side. If this is not the case, an exception will be thrown. The exception contains information
on all the Bean pairs and every mismatching property.

Within BeanMatchStore, there are two routes for a BeanMatch to be created. The main route is the existing one where
a match is set up runtime (or reused if already existing). The second route is the pre-bean instantiation validation
where are registered strict bean pairs are validated upfront. In both cases the validation is done at the time
of BeanMatch construction. The Bean pair validation routine makes use of the underlying exceptions to report them
back all at once.

Configuration allows the setting of the strict mapping convention. If this is enabled, any source class with the
set suffix (default 'Form') will be considered strict. Same counts for any target class with the set suffix
(default 'Result'). This means that the opposing side must have matches for all properties. If these are missing,
this will result in an exception. The principle is basically the same as strict bean pairs, but runtime instead of
before bean instantiation.

Note that strict mapping convention is disabled for dynamically generated classes (downsizeSource/downsizeTarget),
since these concepts rely on the fact that no full match can be supplied.