The name field from an enum is not mapped to target field
Issue #83 In the resolution of issue #78 the definition of getter fields has been tightened, because previously all private fields were tagged as available as well. One project made use of this loophole by reading the name field of an enumeration class to a String field. With the new fix this is no longer possible, since the name field is private.

This fix makes an exception for the name field of an enum class. It will be considered available for reading.