BeanMapper executes the list of AfterClearFlusher instances
Issue #84 states that usage of Hibernate's @OneToMany leads to the known problem of Hibernate first executing its insert, then its delete SQL-statements. When the table uses (as it well should) unique constraints, this leads to a violation. The dirty quickfix is to disable the constraints. The long-term fix is for BeanMapper to force the ORM to flush, effectively reversing the order of delete and insert SQL-statements.

When a clear method on a collection is called, BeanMapper makes sure to run down the list of registered AfterClearFlusher instances. On every instance, the flush method will be called. By default AfterClearFlusher instances are added. BeanMapper has no notion of ORMs; this is left to [beanmapper-spring](https://github.com/42BV/beanmapper-spring) and [beanmapper-spring-boot-starter](https://github.com/42BV/beanmapper-spring-boot-starter).

The calling of the flusher is only executed if the flushAfterClear is explicitly set. If not, it will not call the flusher. It can either be set on the BeanCollection annotation or on the override configuration.