Unmatched BeanProperty did not throw an exception
Issue #60 properties annotated with BeanProperty must match. If they do not, an exception must be thrown. Due to a bug, this did not always occur (only with BeanProperty on the target side). The current mechanism keep tabs on matched properties and does a final verification. If unmatched properties remain that should have been matched, an exception is thrown.

Opened new snapshot version

Added test for #64 to prove it now works