#116 Mappable nested classes analyzed for wrong direction
On dealing with mappable nested classes in combination with a
@BeanProperty dot-delimited route, the wrong mapping direction
was applied. This resulted in errors at the moment that the field
could only be accessed through a method accessor and that the
complementing method accessor was not available (ie, get and no set,
or set and no get).

The problem has been fixed by inverting the method direction on
dealing with a mappable nested class.