#203: Add BeanConverter caching
As some BeanConverters should only be added within the relevant scope,
it turned out to be impossible to create a global store. Rather than
that, we create an instance of the BeanConverterStore only in the
CoreConfiguration, and in the first layer OverrideConfiguration.

As the OverrideConfigurations are stripped from the CoreConfiguration
after a successful mapping, this should guarantee that any cache built
up within OverrideConfigurations is not propagated to the cache within
the CoreConfiguration, while allowing the CoreConfiguration to form its
own cache.

- Updated BeanConverterStore, to make it publicly instantiable.
- Updated CoreConfiguration and OverrideConfiguration, to include the
BeanConverterStore.
- Updated AbstractMapStrategy, to use the BeanConverterStore whenever
possible.