#203: Update BeanMatchStore
Updated BeanMatchStore, to use faster collections, and remove the need
for (relatively) expensive calls to get the names of classes, by simply
passing the classes themselves.

Since classes have an excellent hashCode-implementation, using classes
as the key for the data store, in combination with the relatively fast
HashMap, makes more sense than using the TreeMap.